Compiles, fix warnings on 64 bit
git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@13293 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: ad0320fe62fcc82875050d9fbc2504a6862a39b0

+++ vision/place_recognition/src/py.cpp
@@ -99,7 +99,7 @@ static FeatureMatrix extract_float_features(PyObject *self, PyObject *pim,
 
     // Find keypoints
     detector.DetectPoints(left.Ipl(), std::back_inserter(pts));
-    printf("[Star detector gave %d points, dimension %d]\n", pts.size(), dimension);
+    printf("[Star detector gave %d points, dimension %d]\n", (int)(pts.size()), dimension);
 
     // Compute descriptors
     image_features.resize(pts.size(), dimension);
@@ -160,7 +160,7 @@ static sig_data_t* extract_features(PyObject *self, PyObject *pim,
 
     // Find keypoints
     detector.DetectPoints(left.Ipl(), std::back_inserter(pts));
-    printf("[Star detector gave %d points, dimension %d]\n", pts.size(), dimension);
+    printf("[Star detector gave %d points, dimension %d]\n", (int)(pts.size()), dimension);
 
     // Compute descriptors
     *num_features = pts.size();
